+++ litestar_pagination/__init__.py
@@ -4,4 +4,4 @@ from litestar_pagination.cursor import CursorPage, CursorParams
 
 __all__ = ("CursorPage", "CursorParams")
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"